Limit switch fix

+++ Projects/AVR8Bit/include/util.h
@@ -1,8 +1,8 @@
 #ifndef UTIL_H
 #define UTIL_H
 
-int16_t min(int16_t a, int16_t b);
-int16_t max(int16_t a, int16_t b);
+int32_t min(int32_t a, int32_t b);
+int32_t max(int32_t a, int32_t b);
 uint16_t int_abs(int16_t x);
 void set_LED(uint8_t L, uint8_t state);
 void update_LEDS(uint16_t count);
